feat(boot): 16-to-32-bit boot chain and x86_32 raw kernel target
- Fix codegen32 freestanding _dev_outb/_dev_inb builtins: they read
  [ebp+8] without a prologue, so they scanned the caller's frame and
  issued an IN to a garbage port (froze the vCPU on QEMU's IDE ports).
  Give them a proper cdecl frame (push ebp; mov ebp,esp; ...; leave; ret).
- Have _dev_enter_pmode rep movsd the kernel from the 0x8000 staging
  buffer to 0x100000: SeaBIOS' AH=42h PIO addresses its buffer via a
  16-bit ES:DI pair and cannot write above the low 1 MiB (a 0x100000
  target silently wraps to 0x000000). Drop the now-useless _dev_enable_a20
  builtin.
- Correct the os32 loader: read kernel LBA 3-10 (was 4), into 0x8000
  (was 0xFF00:0x1000 = 0x100F00); fix the DAP layout and CF-check order.
- Add x86_32 FORMAT raw support: LOAD is now a u32 link base driving imm32
  string/global fixups (parser/config/asm/encoder).
- Add examples/os32 (boot sector -> LBA loader -> pmode switch -> kernel),
  examples/targets/x86_32-raw.fld, and the
  os32_boot_chain_reaches_kernel integration test.

All 140 unit / 62 integration tests pass; cargo fmt and clippy are clean.

+++ src/backend/codegen16/mod.rs
@@ -55,7 +55,7 @@ pub(super) use asm::*;
 pub(super) use program::BUILTIN_FUNCS;
 
 impl<'p> CodeGen16<'p> {
-    pub(super) fn new(prog: &'p Program, load_base: u16) -> Self {
+    pub(super) fn new(prog: &'p Program, load_base: u32) -> Self {
         Self {
             prog,
             asm: Encoder::new(load_base),
